+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Next
 
+### Updated
+- Fixed bugs in the Neo4jVector and GraphCypherQAChain classes preventing these classes from working with versions < 5.23 of Neo4j
+
 ## 0.1.0
 
 ### Added